Base Bowler Code In Place
The Base BowlerDevice class, as well as all supporting
classes, are in place.  Additionally, support for
the namespace `bcs.pid` without using introspection
is in place.

Missing are the class for things specific to the DyIO device,
as well as unit tests, and support for higher-level classes
to make using various peripherals easier.  Finally, there is
currently no support for the `bcs.io` namespace, either.