Added Improved Handling for MAC Addresses
Previously, BowlerDevice had loosing handling of the value of MAC
addresses, and didn't have any documentation on how they were supposed
to be formatted.  There are now three values: `_mac_address_raw` (which
is either 'broadcast' or an Array), and two properties: `mac_address`
(which is either 'broadcast' or a ':'-separated string) and
`mac_address_bytes` (which is always an array of bytes).